Bug 1837499 - remove multi_dep loader and related transforms
