+++ taskcluster/ci/signing-apk/kind.yml
@@ -6,7 +6,7 @@ loader: taskgraph.loader.transform:loader
 
 transforms:
     - taskgraph.transforms.from_deps
-    - android_taskgraph.transforms.multi_dep:upstream_artifacts
+    - android_taskgraph.transforms.upstream_artifacts
     - android_taskgraph.transforms.treeherder:transforms
     - android_taskgraph.transforms.signing_apk:transforms
     - android_taskgraph.transforms.notify:transforms
